Create Codex platform
05. Receipt
We've made it to the end of the first chapter! 🎈

console.log("Congrats! You've made it.");

To recap what we've learned:

The console.log() method outputs messages to the Console window.
JavaScript code runs one line at a time, from top to bottom.
Single line comments are created using the // forward slashes.
Multi-line comments are started by /* and closed with */.
It's time to use what we've learned to build something fun!

Instructions
You are going to recreate a receipt of a purchase from the last store or restaurant that you went to... using JavaScript! 🧾

Make sure to check your pockets and wallet if you forgot.

Use console.log() methods to print the following info:

Name of the store/restaurant.
A row that says "Order Details".
A row with column names for "Item", "Quantity", and "Price".
At least two items that were purchased.
The total price for the purchase.
The output should look like:

==================================
         McDONALD'S RECEIPT       
----------------------------------
           Order Details          
----------------------------------
Item           Quantity   Price
----------------------------------
🍔 Burger       1          2.99   
🍟 Fries        1          2.49   
🥤 Soda         1          1.99   
----------------------------------
Total                      7.47   
==================================

When you are done, take a screenshot and click the Twitter icon to share your receipt!

Bonus: Add more details to your receipt like server name and order number.